requirements.txt
requests==2.31.0
sqlite3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
schedule==1.2.0
jinja2==3.1.2
reportlab==4.0.4
smtplib
email-validator==2.0.0
pillow==10.0.0
numpy==1.24.3
datetime
json
logging
time
asyncio
aiohttp==3.8.5